Moved title and added time
Moved title from top to bottom as on the News page the preview is displaying the title, which is just below the main title. Added time so it could accurately display the date of the article.

Need to confirm that the ">" as a quote is rendering well on the website.