Docs for translations: instructions for handling conflicts (#376)
As discussed on Zulip, this aims to outline step-by-step how to handle Weblate conflicts if one accidentally did a force-push to the `website` repo here on Codeberg.

I tried the rebasing locally, and it seems to work like that, but of course didn't run into any real conflicts. @pastk It would be amazing if you could double check if that is indeed all correct based on your first-hand experience!

Signed-off-by: Bastian Greshake Tzovaras <bgreshake@googlemail.com>

Reviewed-on: https://codeberg.org/comaps/website/pulls/376
Reviewed-by: Konstantin Pastbin <pastk@noreply.codeberg.org>
Reviewed-by: map-per <map-per@noreply.codeberg.org>
Co-authored-by: Bastian Greshake Tzovaras <bgreshake@googlemail.com>
Co-committed-by: Bastian Greshake Tzovaras <bgreshake@googlemail.com>